+++ ArmSim/Actuator_Linear.h
@@ -13,13 +13,13 @@ public:
     Actuator_Linear() = delete;
     Actuator_Linear(
         Platform* _ptrPlatform, uint8_t _actuator_ID, std::string _name, ActuatorType _actuatorType,
-        float _posX, float _posY, float _posZ,
-        float _forceVecX, float _forceVecY, float _forceVecZ);
+        double _posX, double _posY, double _posZ,
+        double _forceVecX, double _forceVecY, double _forceVecZ);
 
-    float getRealActuationValue() override;
+    double getRealActuationValue() override;
 
 private:
-    float forceVecX, forceVecY, forceVecZ;
+    double forceVecX, forceVecY, forceVecZ;
 };
 
 #endif
